+++ CHANGELOG.md
@@ -15,6 +15,7 @@ all of them.
  - Add extension `VK_KHR_external_memory_capabilities` (revision 1)
  - Add extension `VK_KHR_external_memory` (revision 1)
  - Add extension `VK_KHR_external_memory_win32` (revision 1)
+ - Add extension `VK_KHR_external_memory_fd` (revision 1)
 
 ### Breaking changes